% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predire_prix.R
\docType{data}
\name{hyundai}
\alias{hyundai}
\title{Hyundai dataset}
\format{
A data frame with 4860 rows and 9 variables:
\describe{
  \item{model}{modele du vehicule}
  \item{year}{annee de fabrication}
  \item{price}{prix du vehicule}
  \item{mileage}{Kilometrage}
  \item{mpg}{Consommation}
  \item{fuelType}{Type de carburant}
  \item{engineSize}{Taille du moteur}
  \item{transmission}{Mode de transmisison du vehicule}
  \item{tax...}{Taxe}
}
}
\source{
\url{https://www.kaggle.com/datasets/mysarahmadbhat/hyundai-used-car-listing}
}
\usage{
hyundai
}
\description{
Un dataset comprenant des observations a propos de vehicules Hyundai d'occasion
}
\examples{
data(hyundai)

}
\keyword{datasets}
