% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_df.R
\name{overview_df}
\alias{overview_df}
\title{Analyse descriptive d'un jeu de donnes}
\usage{
overview_df(donnee)
}
\arguments{
\item{donnee}{Un data frame.}
}
\value{
une liste de 5 elements
\itemize{

  \item \strong{type_variable} : Une matrice qui indique le type de chaque variable du jeu de données (en colonnes). Les types possibles sont :
  \itemize{
    \item \emph{Numerique} : Variables quantitatives.
    \item \emph{Facteur} : Variables qualitatives categoriques.
    \item \emph{Autre} : Variables ne correspondant pas aux types precedents.
  }
  \item \strong{nombre_observations} : Une matrice qui contient, pour chaque variable du jeu de données (en colonnes), les informations suivantes :
  \itemize{
    \item Colonne 1 : Le nombre total d'observations disponibles.
    \item Colonne 2 : Le nombre d'observations manquantes.
  }

  \item \strong{stat_num} : Une matrice contenant les statistiques descriptives des variables numériques du jeu de données (en colonnes). Les statistiques sont presentees en lignes :
  \itemize{
    \item Quantiles empiriques : 0 %, 10 %, 25 %, 50 %, 75 %, 90 %, et 100 %.
    \item Moyenne : La moyenne des valeurs.
    \item Écart-type : La dispersion des valeurs autour de la moyenne.
  }

  \item \strong{stat_fact} : Une liste composée de deux elements decrivant les variables de type facteur :
  \itemize{
    \item \strong{Frequences} : Une liste contenant, pour chaque variable de type facteur, un \code{data.frame} des frequences obtenu via la fonction \code{stat_factor}.
    \item \strong{Mode} : Une liste nommee contenant, pour chaque variable de type facteur, le mode calcule via la fonction \code{stat_factor}.
  }

  \item \strong{var_autre} : Un vecteur contenant les noms des variables classees comme \emph{Autre}.
}
}
\description{
realise une analyse descriptive d'un jeu de donnee.
}
\examples{
data(hyundai)
out_overview <- overview_df(hyundai)
str(out_overview)
}
\author{
Laggoune Yousra, Thierry Fotchou Saadem, Paul Boittiaux
}
