% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageGroupe5-package.R
\docType{package}
\name{PackageGroupe5-package}
\alias{PackageGroupe5}
\alias{PackageGroupe5-package}
\title{PackageGroupe5: Prediction du prix d'un vehicule Hyundai d'occasion}
\description{
Ce package contient les fonctions stats_var_num et stats_fact pour faire des analyses statistique en fonction de la nature d'une variable. Ce package contient aussi une méthode print et une méthode summary pour chacun de ces 2 fonctions. Le package contient aussi une fonction overview_df qui permet d'avoir un analyse statistique détaille d'un dataframe. Cette fonction à aussi une fonction summary intégré dans le package. Il y a enfin la fonction predire_prix qui permet de predire le prix d'un vehicule d'occasion hyundai a partir de ses caracteristiques. Que ce soit pour de la vente ou un achat, les utilisateurs peuvent se fier a ce package pour les aider au travers de cette etape. Un modele lineaire comprenant 7 variables est utilise pour faire les predictions, ce le rend solide et fiable.
}
\keyword{internal}
