% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predire_prix.R
\name{predire_prix}
\alias{predire_prix}
\alias{print.prix_estime}
\title{Prediction du prix}
\usage{
predire_prix(
  year,
  mileage,
  transmission,
  model,
  fuelType,
  mpg,
  engineSize,
  data
)

\method{print}{prix_estime}(x, ...)
}
\arguments{
\item{year}{Annee de construction du vehicule. Prediction plus precise pour les vehicules produits apres 2000.}

\item{mileage}{Kilometrage du vehicule, avec un maximum recommande de 250 000km.}

\item{transmission}{Type de transmission, au choix parmi "Automatic", "Semi-Auto" et "Manual".}

\item{model}{Modele du vehicule parmi les modeles suivants:"I20","Tucson","I10","IX35","I30","I40","Ioniq", "Kona","Veloster","I800","IX20","Santa Fe","Accent","Terracan","Getz","Amica".}

\item{fuelType}{Type de carburant, au choix entre "Diesel", "Hybrid", "Petrol" ou "Other".}

\item{mpg}{Consommation du vehicule.}

\item{engineSize}{Taille du moteur du vehicule.}

\item{data}{Dataset contenant les informations pour entrainer le modele}

\item{x}{Un objet produit par la fonction \code{predire_prix}, a afficher}

\item{\dots}{D'autres arguments passes a d'autre methodes}
}
\value{
\code{"prix_estime"}: un objet de classe "prix_estime" qui represente le prix estime par le modele
}
\description{
Prediction du prix d'un vehicule Hyundai d'occasion selon ses caracteristiques.

Cette fonction predit le prix d'un vehicule base sur divers parametres.
}
\section{Functions}{
\itemize{
\item \code{print(prix_estime)}: Affiche un objet de la classe \code{"prix_estime"}.

}}
\examples{
data(hyundai)
prix_prediction <- predire_prix(year=2020,mileage=10000,
                     transmission="Manual",model="I20",fuelType ="Hybrid",
                     mpg=45,engineSize=1.8,data=hyundai)
print(prix_prediction)
print(prix_prediction)

}
\author{
Laggoune Yousra, Thierry Fotchou Saadem, Paul Boittiaux
}
