% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_print_stats_fact.R
\name{print.stats_fact}
\alias{print.stats_fact}
\title{Methode print pour statistique de variable categorique}
\usage{
\method{print}{stats_fact}(x, ...)
}
\arguments{
\item{x}{Un objet de classe stats_fact retourne par la fonction stats_fact.}

\item{...}{Arguments supplementaires}
}
\value{
Cette methode imprime :
\itemize{
  \item Un tableau des frequences (modalites, frequences absolues, proportions).
 }
}
\description{
Cette methode pour la fonction generique print affiche un resume detaille des statistiques calculees par la fonction stats_fact.
}
\examples{
data(hyundai)
sortie_b <- stats_fact(hyundai$model)
print(sortie_b)
}
\author{
Laggoune Yousra, Thierry Fotchou Saadem, Paul Boittiaux
}
