% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_print_stats_var_num.R
\name{print.stats_var_num}
\alias{print.stats_var_num}
\title{Methode print pour statistiques numérique}
\usage{
\method{print}{stats_var_num}(x, ...)
}
\arguments{
\item{x}{Un objet de classe stats_var_num retourne par la fonction stats_var_num}

\item{...}{Arguments supplementaires passes a la fonction print}
}
\value{
Cette methode imprime un extrait de l'objet contenant:

\itemize{
  \item Les quantiles
  \item La moyenne
  \item L'ecart-type   }
}
\description{
Cette methode pour la fonction generique print affiche un resume des statistiques
calculees par la fonction stats_var_num
}
\examples{
data(hyundai)
sortie_a <- stats_var_num(hyundai$mpg)
print(sortie_a)
}
\author{
Laggoune Yousra, Thierry Fotchou Saadem, Paul Boittiaux
}
