% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_fact.R
\name{stats_fact}
\alias{stats_fact}
\title{Statistique pour variable categorique}
\usage{
stats_fact(vect_fact, ...)
}
\arguments{
\item{vect_fact}{Un vecteur contenant les observations de la variable a traiter.}

\item{...}{Arguments supplementaires}
}
\value{
Une liste de cinq elements :
\itemize{
  \item \strong{nombre_observations} : Un vecteur contenant deux valeurs :
    \itemize{
      \item Le nombre total d'observations.
      \item Le nombre d'observations manquantes (NA).
    }
  \item \strong{frequences} : Un data frame contenant :
    \itemize{
      \item \strong{Modalite} : Les modalites de la variable (excluant les valeurs manquantes).
      \item \strong{Frequence} : Les frequences des modalites.
      \item \strong{Proportion} : Les proportions relatives des modalites.
    }
  \item \strong{mode} : Le mode de la variable (la modalite la plus frequente).
  \item \strong{observations} : Une copie du vecteur d'observations fourni.
}
}
\description{
La fonction \code{stats_fact} sert a decrire les observations d'une variable categorielle. Elle calcule les statistiques de base, y compris les frequences, les proportions et le mode.
}
\examples{
sortie_b <- stats_fact(hyundai$model)
str(sortie_b)
}
\author{
Laggoune Yousra, Thierry Fotchou Saadem, Paul Boittiaux
}
