% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_var_num.R
\name{stats_var_num}
\alias{stats_var_num}
\title{Statistiques pour une variable numerique}
\usage{
stats_var_num(vect_num, ...)
}
\arguments{
\item{vect_num}{Un vecteur numerique contenant les observations de la variable a traiter.}

\item{...}{Argument supplementaires}
}
\value{
Une liste de cinq éléments :
\itemize{
  \item \strong{nombre_observations} : Un vecteur nommé contenant deux valeurs :
  \item \strong{quantiles} : Un vecteur contenant les quantiles suivants :
    0%, 10%, 25%, 50%, 75%, 90%, 100%.
  \item \strong{moyenne} : La moyenne des observations.
  \item \strong{ecarttype} : L'ecart-type des observations.
  \item \strong{observations} : Un vecteur numérique contenant les valeurs des observations,
    y compris les observations manquantes.
}
}
\description{
La fonction \code{stats_num} sert a calculer des statistiques decrivant les observations d'une variable numerique.
}
\examples{
data(hyundai)
sortie_a <- stats_var_num(hyundai$mpg)
str(sortie_a)
}
\author{
Laggoune Yousra, Thierry Fotchou Saadem, Paul Boittiaux
}
