% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_overview_df.R
\name{summary.overview_df}
\alias{summary.overview_df}
\title{Methode summary pour analyse descriptive d'un jeu de donnees}
\usage{
\method{summary}{overview_df}(object, ...)
}
\arguments{
\item{object}{Un objet genere par la fonction \code{overview} qui doit etre affiche.}

\item{\dots}{Arguments supplementaires transmis a dautres methodes.}
}
\description{
Cette fonction est la methode S3 pour la fonction d'overview qui permet d'afficher ce que l'on veut en fonction de l'entree
}
\examples{
data(hyundai)
object <- overview_df(hyundai)
summary(object)
}
\author{
Laggoune Yousra, Thierry Fotchou Saadem, Paul Boittiaux
}
