% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_summary_stats_fact.R
\name{summary.stats_fact}
\alias{summary.stats_fact}
\title{Methode summary pour statistique de variable categorique}
\usage{
\method{summary}{stats_fact}(object, ...)
}
\arguments{
\item{object}{Un objet de classe stats_fact retourne par la fonction stats_fact.}

\item{...}{Arguments supplementaires}
}
\value{
Cette methode affiche:
\itemize{
  \item Le nombre total d'observations et les observations manquantes.
  \item Un tableau des frequences (modalites et frequences absolues).
  \item Le mode (la categorie la plus frequente).
}
}
\description{
Cette methode pour la fonction generique summary affiche un resume des statistiques
calculees par la fonction stats_fact
}
\examples{
data(hyundai)
sortie_b <- stats_fact(hyundai$model)
summary(sortie_b)
}
\author{
Laggoune Yousra, Thierry Fotchou Saadem, Paul Boittiaux
}
