% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_summary_stats_var_num.R
\name{summary.stats_var_num}
\alias{summary.stats_var_num}
\title{Methode summary pour statistique numérique}
\usage{
\method{summary}{stats_var_num}(object, ...)
}
\arguments{
\item{object}{Un objet de classe \code{stats_var_num} retourne par la fonction \code{stats_var_num}.}

\item{...}{Arguments supplementaires passes a d'autres methodes (non utilises ici).}
}
\value{
Cette methode imprime un resume contenant :
\itemize{
  \item Le nombre total d'observations et le nombre d'observations manquantes.
  \item Les quantiles.
  \item La moyenne et l'ecart-type.
  }
}
\description{
Cette methode pour la fonction generique affiche un resume detaille des
statistiques d un objet de classe \code{stats_var_num} retourne par la fonction \code{stats_var_num}.
}
\examples{
sortie_a <- stats_var_num(hyundai$mpg)
print(sortie_a)
}
\author{
Laggoune Yousra, Thierry Fotchou Saadem, Paul Boittiaux
}
